requirements.txt
Django==1.9.8
django-annoying==0.9.0
python-dateutil==2.5.3
django-daterange-filter==1.3.0
